Project/requirement.txt
streamlit
ultralytics
opencv-python
numpy
Pillow
sahi
torch
ensemble_boxes